% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rates.fit}
\alias{rates.fit}
\title{Fit the RATES model.}
\usage{
rates.fit(
  betajk,
  sjk2,
  PC,
  p = 0.003,
  lambda = 0.04,
  tau2 = NULL,
  alpha = 3,
  SNP = NULL,
  rel.eps = 1e-08,
  verbose = 1,
  maxIter = 10^4L,
  minIter = 50
)
}
\arguments{
\item{betajk}{MxK matrix of effect size estimates, where row j corresponds to j-th SNP, and column k corresponds to k-th study .  Missing values can be represented with NA.}

\item{sjk2}{MxK matrix of effect size estimate variances, where row j corresponds to SNP j, and column k corresponds to study k.  Missing values can be represented with NA.}

\item{PC}{Kx4 matrix of PCs (PC0,PC1,PC2,PC3) of genome-wide AF for K studies.}

\item{p}{initial value for EM algorithm, for the proportion of non-zero SNPs}

\item{lambda}{initial value for EM algorithm, for the proportion (P(O_jk=1|Rj=0)) of non-replicable SNPs which are outliers with inflated variance.}

\item{tau2}{initial value for EM algorithm, a (L+1)-vector for the variance of coefficients of PCs (gamma_jl,l= 0,1,2,3,if L=3) for replicable non-zero effect SNPs.}

\item{alpha}{initial value for EM algorithm, for the variance inflation of outlier summary statistics at non-replicable SNPs.}

\item{rel.eps}{threshold for when to end EM algorithm. rel.eps = (ll\link{i} - ll\link{i-1})/ll\link{i}, where ll\link{i} is the log-likelihood at iteration i.}

\item{maxIter}{maximum # of EM iterations.}

\item{minIter}{minimum # of EM iterations.}
}
\value{
A list.
}
\description{
Fit the RATES model.
}
