% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rates.generative}
\alias{rates.generative}
\title{RATES generative model}
\usage{
rates.generative(mod, sjk2, seed = NULL, Mnull = NULL)
}
\arguments{
\item{mod}{a fitted rates model}

\item{sjk2}{a M x K matrix of effect size variance used for fitting the original rates model/for rates generative model}

\item{Mnull}{Approximate number of null variants to generate}
}
\value{
a list of generated data
}
\description{
RATES generative model
}
