% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rates.getnullscore}
\alias{rates.getnullscore}
\title{parametric boostrap using RATES generative model}
\usage{
rates.getnullscore(
  mod,
  sjk2,
  total_null_scores,
  out.prefix,
  seed = NULL,
  nullSNPsPerModel = NULL,
  clean.existing.score = F
)
}
\arguments{
\item{mod}{a fitted rates model}

\item{sjk2}{a M x K matrix of effect size variance used for fitting the original rates model}

\item{seed}{seed used in bootstrap sampling}

\item{nullSNPsPerModel}{apprixamet number of null variants generated in each fold of parametric bootstrap}

\item{total_null_score}{total number of ppr of simulated null SNPs}
}
\value{
a list of null scores and used seed.
}
\description{
parametric boostrap using RATES generative model
}
